Move docstring handling to _utils
These functions are general enough that they will be needed by many
wrappers. So it makes more sense to refactor them into a common private
module and use them from there.
